fix: crash when creating new_window (ctrl+shift+n)
Fix a crash when creating a new window using the default ctrl+shift+n
keybind. This keybind is intended to create a new window with the
current working directory.

The crash first occurs because data_device and selection_device can only
have one listener. When we init a new surface, we would add a new
listener and would crash (note: the crash doesn't occur in ReleaseFast
here, but in Debug it does). Move the listener to the Seat struct and
act on the active surface.

The second failure mode is in the keyboardListener. We don't update the
surface of the Seat when a surface is deinited, so the next keyboard
event results in a segfault because the surface is still on the seat
variable, but has been deinited. Set the seat surface to null on surface
deinit to resolve this.