input: refactor key repeat
The timer doesn't need to be an optional, it is always around. Prefer to
use `reset` over cancel + flags to monitor if the timer is running.
Reset cancels the timer if it is already running, which is exactly what
we want to do when we need to restart the timer for a new key repeat.
This also prevents needing an explicit cancel prior to starting the
timer.

We can avoid cancelling the timer another way by only setting the
`pressed_key` field when we have a key which may need repeating. In all
other cases we set it to null. When the timer fires, it will self disarm
because the field is set to null.