app: run app wakeup tick separately from display dispatch
Terminal applications can receive certain actions in the background
which require the app to tick. Currently, we block on display.dispatch
before ticking the app. However if the terminal window is not visible,
this will block until the display *is* visible, after which point we
will tick pending actions.

One example of this is a desktop notification from a running terminal
app. To see how this works, run:

  sleep 3 && printf "\x1b]777;notify;hi;hey\x1b\\"

While the sleep is in effect, switch to another workspace so the
terminal is not visible. The notification will not appear until the
terminal is displayed again.

To fix this, we add a new callback for the app.wakeup completion. This
callback ticks the app, but does not handle anything related to the
wayland Display.